Metrics and Categories
Justin Tomlinson

I have implemented the postgresql database (which will not be in GitHub due
to the large file size), queries for most of the metrics, the Gephi categories,
queries for most of the metrics per category, text files containing the
results of both queries, and a README to describe it all.

I still need to finish some metrics, do some correlation tests,
make some graphs, and look into the ten fastest users to 20,000 reputation
in stack overflow, and possibly draw ome different conclusions.