Added new category query example for finding "experts"
Justin Tomlinson

The categoriesquery.sql file now contains a query to find the
experts of a tag. It's pretty basic and more or less just narrows it
down, instead of giving you an exact result of who is an expert. It
would probably pretty easy to modify it to specify it for the different
kinds of experts though.

I have also made a small change to the README to indicate that the
queries in the categoriesquery.sql file will not be exactly the same
as the results in the paper as they are for demonstration and I did not
actually use these scripts in the database. There are a lot more tags per
category, so I just included the top tags as listed in the paper.